Project 2
Based on Lab 2, the project allows an infinite amount of grade entries with names to follow, and will use the average of all grades to assign a letter grade to all entries. Can be used to modify an existing file already if only partial progress has been made. Will also over-ride a previous grade to give the new grade if applicable